+++ CHANGES.md
@@ -2,15 +2,27 @@
 
 ## [Unreleased]
 
-- Fix Docker compose file, so example data can be loaded into database (author @zstatmanweil, <https://github.com/stac-utils/stac-fastapi-pgstac/pull/142>)
+### Changed
+
 - Handle `next` and `dev` tokens now returned as links from pgstac>=0.9.0 (author @zstatmanweil, https://github.com/stac-utils/stac-fastapi-pgstac/pull/140)
 - Add collection search extension ([#139](https://github.com/stac-utils/stac-fastapi-pgstac/pull/139))
 - keep `/search` and `/collections` extensions separate ([#158](https://github.com/stac-utils/stac-fastapi-pgstac/pull/158))
-- Fix `filter` extension implementation in `CoreCrudClient`
 - update `pypgstac` requirement to `>=0.8,<0.10`
 - set `pypgstac==0.9.*` for test requirements
+- update `stac-fastapi-*` requirement to `~=4.0`
+- remove `python 3.8` support
+- renamed `post_request_model` attribute to `pgstac_search_model` in `CoreCrudClient` class
+- changed `datetime` input type to `string` in GET endpoint methods
+- renamed `filter` to `filter_expr` input attributes in GET endpoint methods
+- delete `utils.format_datetime_range` function
+
+### Fixed
+
+- Fix Docker compose file, so example data can be loaded into database (author @zstatmanweil, <https://github.com/stac-utils/stac-fastapi-pgstac/pull/142>)
+- Fix `filter` extension implementation in `CoreCrudClient`
 
 ## [3.0.1] - 2024-11-14
+
 - Enable runtime `CORS` configuration using environment variables (`CORS_ORIGIN="https://...,https://..."`, `CORS_METHODS="PUT,OPTIONS"`) (https://github.com/stac-utils/stac-fastapi-pgstac/pull/168)
 
 ## [3.0.0] - 2024-08-02